pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csao-reco"
version = "0.1.0"
description = "Cart Super Add-On recommendation scaffold"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.3",
  "joblib>=1.3",
  "lightgbm>=4.0",
]

[tool.setuptools]
packages = [
  "common",
  "data",
  "features",
  "candidate_generation",
  "candidate_generation.retrievers",
  "candidate_generation.rules",
  "ranking",
  "ranking.training",
  "ranking.inference",
  "evaluation",
  "evaluation.metrics",
  "evaluation.segments",
  "serving",
  "serving.api",
  "serving.pipeline",
  "serving.utils",
]
